practices/requirements.txt
art
argparse
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic

langchain
langchain-core
langchain-openai
langchain-experimental
langchain-community

langchain-pinecone
pinecone-client
wikipedia

faiss-cpu
arxiv

rank_bm25

chromadb
langchain-chroma